Tests/LibWeb/Text/expected/abspos-relayout-container-query.txt

color-before=rgb(0, 0, 0)
container=10,10,200,60
color-after=rgb(0, 128, 0)
partial-layouts=1 full-layouts=0
PASS
